refactor: :truck: ディレクトリ名を変更
