+++ lib/feature/word/presentation/component/index_tile.dart
