+++ lib/feature/word/presentation/page/index_second/index_second_page.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import '../../../../../core/common_widget/button/back_icon_button.dart';
 import '../../../../../core/router/app_router.dart';
 import '../../../util/initial_main_group.dart';
-import '../component/index_tile.dart';
+import '../../component/index_tile.dart';
 
 @RoutePage()
 class IndexSecondPage extends StatelessWidget {
