+++ lib/feature/word/presentation/page/search_word_result/search_word_result_page.dart
@@ -3,13 +3,13 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../../../../../core/common_widget/button/back_icon_button.dart';
-import '../../../../../core/common_widget/cupertino_refresh_indicator.dart';
-import '../../../../../core/common_widget/infinite_scroll_bottom_indicator.dart';
-import '../../../../../util/logger.dart';
-import '../../application/word_list_state_by_search_word.dart';
-import 'component/word_tile.dart';
-import 'index_top_page/word_search_text_field.dart';
+import '../../../../../../core/common_widget/button/back_icon_button.dart';
+import '../../../../../../core/common_widget/cupertino_refresh_indicator.dart';
+import '../../../../../../core/common_widget/infinite_scroll_bottom_indicator.dart';
+import '../../../../../../util/logger.dart';
+import '../../../application/word_list_state_by_search_word.dart';
+import '../../component/word_tile.dart';
+import '../index_top/word_search_text_field.dart';
 
 @RoutePage()
 class SearchWordResultPage extends ConsumerWidget {
